Add files via upload
In this file, you will find the code used to perform learning on MNIST using differential evolution-based CNN. This code is also enabling CPU-GPU hybrid acceleration method to speed up the learning. For the CPU part, we used the MPI library, and for GPU the PyTorch built-in functions. To use this code on any other dataset or problem, you only have to change the architecture of the neural network and the function dataset(). (pay attention to the variable containing the number of layers.).